+++ src/api/promo.ts
@@ -1,5 +1,5 @@
 import { getRequestToken } from '@/utils/request-token-handler';
-import { fetchFromApi } from './platformApi';
+import { fetchFromApi } from './platform-api';
 import type {
   CartDiscountPagedQueryResponse,
   CartUpdateAction,
