+++ src/components/product-card/product-card.tsx
@@ -12,8 +12,8 @@ import { useNavigate } from 'react-router';
 import { ROUTES } from '@/router/routes';
 import { PriceBlock } from '../price-block/price-block';
 import Button from '@mui/material/Button';
-import { addProductToCart } from '@/api/addProductToCart';
-import { removeProductFromCart } from '@/api/removeProductFromCart';
+import { addProductToCart } from '@/api/add-product-to-cart';
+import { removeProductFromCart } from '@/api/remove-product-from-cart';
 import { useEffect, useState } from 'react';
 import { useToast } from '@/context/toast-provider';
 import { useCart } from '@/context/cart-context';
