+++ src/components/quantity-input/quantity-input.tsx
@@ -2,7 +2,7 @@ import Button from '@mui/material/Button';
 import styles from './quantity-input.module.scss';
 import { useMutation } from '@tanstack/react-query';
 import { getRequestToken } from '@/utils/request-token-handler';
-import { changeCardItemQuantity } from '@/api/changeCardItemQuantity';
+import { changeCardItemQuantity } from '@/api/change-card-item-quantity';
 import { useCart } from '@/context/cart-context';
 import { useToast } from '@/context/toast-provider';
 import { ERROR_MESSAGES, SUCCESS_MESSAGES } from '@/utils/constants/messages';
