+++ src/hooks/use-customer-query.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { useToken } from '@/context/token-context';
-import { fetchFromApi } from '@/api/platformApi';
+import { fetchFromApi } from '@/api/platform-api';
 import type { Customer } from '@commercetools/platform-sdk';
 
 export const useCustomerQuery = () => {
