+++ src/pages/registration-page/registration-page.tsx
@@ -4,8 +4,8 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { registrationSchema } from '@/validation/registration-validation';
 import type { Addresses, RegisterFormInputs } from '@/validation/registration-validation';
 import styles from './registration-page.module.scss';
-import { getCustomerToken } from '@/api/platformApi';
-import { registerCustomer } from '@/api/clientAuth';
+import { getCustomerToken } from '@/api/platform-api';
+import { registerCustomer } from '@/api/client-auth';
 import { useAuth } from '@/hooks/use-auth';
 import { ERROR_MESSAGES, SUCCESS_MESSAGES } from '@/utils/constants/messages';
 import { useToast } from '@/context/toast-provider';
